[Go SDK] Use the known varint64 coder for int64s.

+++ sdks/go/pkg/beam/core/graph/coder/coder.go
@@ -246,7 +246,7 @@ func NewBytes() *Coder {
 
 // NewVarInt returns a new int32 coder using the built-in scheme.
 func NewVarInt() *Coder {
-	return &Coder{Kind: VarInt, T: typex.New(reflectx.Int32)}
+	return &Coder{Kind: VarInt, T: typex.New(reflectx.Int64)}
 }
 
 // IsW returns true iff the coder is for a WindowedValue.
